notifications enabled

+++ MunicipaliOmegaDatabase/src/main/java/acropollis/municipali/omega/database/db/service/article/ArticleServiceImpl.java
@@ -96,7 +96,9 @@ public class ArticleServiceImpl implements ArticleService {
         articleModel.getQuestions().clear();
         articleModel.getTranslatedArticles().clear();
 
-        articleToReleasePushRecordDao.delete(articleId);
+        if (articleToReleasePushRecordDao.exists(articleId)) {
+            articleToReleasePushRecordDao.delete(articleId);
+        }
     }
 
     private void clearIcons(ArticleModel article) {
